#113 chore : 분석 패키지 분리
