+++ src/main/java/com/floney/floney/book/controller/BookController.java
@@ -132,19 +132,4 @@ public class BookController {
         return new ResponseEntity<>(bookService.inviteCode(bookKey), HttpStatus.OK);
     }
 
-    @GetMapping("/analyze/category")
-    public ResponseEntity<?> analyzeByCategory(@RequestBody AnalyzeByCategoryRequest request) {
-        return new ResponseEntity<>(bookService.analyzeByCategory(request), HttpStatus.OK);
-    }
-
-    @PostMapping("/analyze/budget")
-    public ResponseEntity<?> analyzeByBudget(@RequestBody AnalyzeRequestByBudget request) {
-        return new ResponseEntity<>(bookService.analyzeByBudget(request), HttpStatus.OK);
-    }
-
-    @PostMapping("/analyze/asset")
-    public ResponseEntity<?> analyzeByAsset(@RequestBody AnalyzeRequestByAsset request) {
-        return new ResponseEntity<>(bookService.analyzeByAsset(request), HttpStatus.OK);
-    }
-
 }
